[Chore] app 전용 타겟 메이커 func 추가, 각 타입 클래스 분리
