+++ juinjang/Projects/App/Project.swift
@@ -1,8 +1,30 @@
 import ProjectDescription
 import ProjectDescriptionHelpers
 
-let debugConfig = Path.relativeToRoot("Projects/XCConfig/App/Debug.xcconfig")
-let releaseConfig = Path.relativeToRoot("Projects/XCConfig/App/Release.xcconfig")
+let project = Project(
+    name: "App",
+    targets: [
+        // MARK: - PROD
+        .makeAppTarget(
+            appType: .prod,
+            appDependencies: appDependencies,
+            appVersion: "2.0.2",
+            build: "2025.08.25.1"
+        ),
+        
+        // MARK: - DEV
+        .makeAppTarget(
+            appType: .dev,
+            appDependencies: appDependencies,
+            appVersion: "2.0.2",
+            build: "2025.08.25.1"
+        )
+    ],
+    additionalFiles: [
+        .glob(pattern: .plist.googleServiceInfoDebug),
+        .folderReference(path: .config.sharedConfig)
+    ]
+)
 
 let appDependencies: [Module] = [
     .core(.common),
@@ -15,8 +37,6 @@ let appDependencies: [Module] = [
     .domain(.services),
     .domain(.usecases),
     .domain(.usecaseInterfaces),
-    .designSystem(.resources),
-    .designSystem(.components),
     .presentation(.scenes),
     .spm(.alamofire),
     .spm(.amplitude),
@@ -42,71 +62,3 @@ let appDependencies: [Module] = [
     .spm(.toast),
     .spm(.firebaseFirestore)
 ]
-
-let project = Project(
-    name: "App",
-    targets: [
-        
-        // MARK: - PROD
-        .make(
-            name: "juinjang",
-            product: .app,
-            bundleId: "com.juinjangteam.juinjang",
-            infoPlist: .file(path: .plist.appInfo),
-            sources: ["Sources/**"],
-            resources: ["Resources/**", "Sources/Manager/InAppPurchase/Products.storekit"],
-            entitlements: .file(path: .relativeToRoot("Projects/App/Entitlements/juinjang.entitlements")),
-            dependencies: appDependencies,
-            settings: .settings(
-                base: [
-                    "ASSETCATALOG_COMPILER_GENERATE_SWIFT_ASSET_SYMBOL_EXTENSIONS": "YES",
-                    "OTHER_LDFLAGS": "-ObjC",
-                    "ASSETCATALOG_COMPILER_APPICON_NAME": "AppIcon",
-                    "CFBundleDisplayName": "주인장",
-                    "SWIFT_ACTIVE_COMPILATION_CONDITIONS": "PROD",
-                    "MARKETING_VERSION": "2.0.2", // Version
-                    "CURRENT_PROJECT_VERSION": "2025.08.25.1" // Build
-                ],
-                configurations: [
-                    .debug(name: .debug, xcconfig: releaseConfig),
-                    .release(name: .release, xcconfig: releaseConfig)
-                ])
-        ),
-        
-        // MARK: - DEV
-        .make(
-            name: "juinjang-dev",
-            product: .app,
-            bundleId: "com.juinjangteam.juinjang",
-            infoPlist: .file(path: .plist.appInfoDev),
-            sources: ["Sources/**"],
-            resources: ["Resources/**", "Sources/Manager/InAppPurchase/Products.storekit"],
-            entitlements: .file(path: .relativeToRoot("Projects/App/Entitlements/juinjang.entitlements")),
-            dependencies: appDependencies,
-            settings: .settings(
-                base: [
-                    "ASSETCATALOG_COMPILER_GENERATE_SWIFT_ASSET_SYMBOL_EXTENSIONS": "YES",
-                    "OTHER_LDFLAGS": "-ObjC",
-                    "ASSETCATALOG_COMPILER_APPICON_NAME": "AppIcon-dev",
-                    "CFBundleDisplayName": "주인장-개발",
-                    "SWIFT_ACTIVE_COMPILATION_CONDITIONS": "DEV",
-                    "MARKETING_VERSION": "2.0.2", // Version
-                    "CURRENT_PROJECT_VERSION": "2025.08.25.1" // Build
-                ],
-                configurations: [
-                    .debug(name: .debug, xcconfig: debugConfig),
-                    .release(name: .release, xcconfig: debugConfig)	
-                ])
-        )
-    ],
-    additionalFiles: [
-        .glob(pattern: .plist.googleServiceInfoDebug),
-        .folderReference(path: .config.sharedConfig)
-    ]
-)
-
-public extension TargetDependency {
-    static func spm(_ name: String) -> TargetDependency {
-        return .external(name: name)
-    }
-}
