+++ juinjang/Tuist/ProjectDescriptionHelpers/SPM.swift
@@ -1,0 +1,35 @@
+//
+//  SPM.swift
+//  Manifests
+//
+//  Created by KimDongWoo on 8/29/25.
+//
+
+// MARK: - SPM
+public enum SPM: String {
+    case alamofire = "Alamofire"
+    case amplitude = "AmplitudeSwift"
+    case dgCharts = "DGCharts"
+    case dsWaveformImage = "DSWaveformImage"
+    case dsWaveformImageViews = "DSWaveformImageViews"
+    case firebaseAnalytics = "FirebaseAnalytics"
+    case firebaseFirestore = "FirebaseFirestore"
+    case fsCalendar = "FSCalendar"
+    case iqKeyboardManager = "IQKeyboardManagerSwift"
+    case kakaoCommon = "KakaoSDKCommon"
+    case kakaoAuth = "KakaoSDKAuth"
+    case kakaoShare = "KakaoSDKShare"
+    case kakaoUser = "KakaoSDKUser"
+    case kingfisher = "Kingfisher"
+    case lottie = "Lottie"
+    case reactorKit = "ReactorKit"
+    case realmSwift = "RealmSwift"
+    case rxSwift = "RxSwift"
+    case rxCocoa = "RxCocoa"
+    case rxDataSources = "RxDataSources"
+    case skeletonView = "SkeletonView"
+    case snapKit = "SnapKit"
+    case tabman = "Tabman"
+    case then = "Then"
+    case toast = "Toast"
+}
